refactor: project users virtual autocomplete
