+++ frontend/src/component/project/ProjectAccess/ProjectAccessAssign/ProjectAccessAssign.tsx
@@ -40,6 +40,7 @@ import { MultipleRoleSelect } from 'component/common/MultipleRoleSelect/Multiple
 import type { IUserProjectRole } from '../../../../interfaces/userProjectRoles';
 import { useCheckProjectPermissions } from 'hooks/useHasAccess';
 import { ADMIN } from 'component/providers/AccessProvider/permissions';
+import AutocompleteVirtual from '../../../common/AutocompleteVirtual/AutcompleteVirtual';
 
 const StyledForm = styled('form')(() => ({
     display: 'flex',
@@ -339,6 +340,9 @@ export const ProjectAccessAssign = ({
             userRoles.some((userrole) => role.id === userrole.id),
         );
     }
+
+    const isLargeList = options.length > 200;
+
     return (
         <SidebarModal
             open
@@ -362,101 +366,206 @@ export const ProjectAccessAssign = ({
                             Select the {entityType}
                         </StyledInputDescription>
                         <StyledAutocompleteWrapper>
-                            <Autocomplete
-                                data-testid={PA_USERS_GROUPS_ID}
-                                size='small'
-                                multiple
-                                openOnFocus
-                                limitTags={10}
-                                disableCloseOnSelect
-                                disabled={edit}
-                                value={selectedOptions}
-                                getOptionDisabled={(option) => {
-                                    if (option.type === ENTITY_TYPE.GROUP) {
-                                        const optionGroup =
-                                            option.entity as IGroup;
-                                        return Boolean(optionGroup.rootRole);
-                                    }
-                                    return false;
-                                }}
-                                onChange={(event, newValue, reason) => {
-                                    if (
-                                        event.type === 'keydown' &&
-                                        (event as React.KeyboardEvent).key ===
-                                            'Backspace' &&
-                                        reason === 'removeOption'
-                                    ) {
-                                        return;
-                                    }
-                                    setSelectedOptions(newValue);
-                                }}
-                                options={options}
-                                groupBy={(option) => option.type}
-                                renderOption={(props, option, { selected }) =>
-                                    renderOption(props, option, selected)
-                                }
-                                getOptionLabel={(option: IAccessOption) => {
-                                    if (
-                                        option.type === ENTITY_TYPE.USER ||
-                                        option.type ===
-                                            ENTITY_TYPE.SERVICE_ACCOUNT
-                                    ) {
-                                        const optionUser =
-                                            option.entity as IUser;
-                                        return (
-                                            optionUser.email ||
-                                            optionUser.name ||
-                                            optionUser.username ||
-                                            ''
-                                        );
-                                    } else {
-                                        return option.entity.name;
-                                    }
-                                }}
-                                filterOptions={(options, { inputValue }) =>
-                                    options
-                                        .filter((option: IAccessOption) => {
-                                            if (
-                                                option.type ===
-                                                    ENTITY_TYPE.USER ||
-                                                option.type ===
-                                                    ENTITY_TYPE.SERVICE_ACCOUNT
-                                            ) {
-                                                const optionUser =
-                                                    option.entity as IUser;
-                                                return (
-                                                    caseInsensitiveSearch(
-                                                        inputValue,
-                                                        optionUser.email,
-                                                    ) ||
-                                                    caseInsensitiveSearch(
-                                                        inputValue,
-                                                        optionUser.name,
-                                                    ) ||
-                                                    caseInsensitiveSearch(
-                                                        inputValue,
-                                                        optionUser.username,
-                                                    )
-                                                );
-                                            }
-                                            return caseInsensitiveSearch(
-                                                inputValue,
-                                                option.entity.name,
+                            {isLargeList ? (
+                                <AutocompleteVirtual
+                                    data-testid={PA_USERS_GROUPS_ID}
+                                    size='small'
+                                    multiple
+                                    openOnFocus
+                                    limitTags={10}
+                                    disableCloseOnSelect
+                                    disabled={edit}
+                                    value={selectedOptions}
+                                    getOptionDisabled={(option) => {
+                                        if (option.type === ENTITY_TYPE.GROUP) {
+                                            const optionGroup =
+                                                option.entity as IGroup;
+                                            return Boolean(
+                                                optionGroup.rootRole,
                                             );
-                                        })
-                                        .slice(0, 100)
-                                }
-                                isOptionEqualToValue={(option, value) =>
-                                    option.type === value.type &&
-                                    option.entity.id === value.entity.id
-                                }
-                                renderInput={(params) => (
-                                    <TextField
-                                        {...params}
-                                        label={capitalize(entityType)}
-                                    />
-                                )}
-                            />
+                                        }
+                                        return false;
+                                    }}
+                                    onChange={(event, newValue, reason) => {
+                                        if (
+                                            event.type === 'keydown' &&
+                                            (event as React.KeyboardEvent)
+                                                .key === 'Backspace' &&
+                                            reason === 'removeOption'
+                                        ) {
+                                            return;
+                                        }
+                                        setSelectedOptions(newValue);
+                                    }}
+                                    options={options}
+                                    renderOption={(
+                                        props,
+                                        option,
+                                        { selected },
+                                    ) => renderOption(props, option, selected)}
+                                    getOptionLabel={(option: IAccessOption) => {
+                                        if (
+                                            option.type === ENTITY_TYPE.USER ||
+                                            option.type ===
+                                                ENTITY_TYPE.SERVICE_ACCOUNT
+                                        ) {
+                                            const optionUser =
+                                                option.entity as IUser;
+                                            return (
+                                                optionUser.email ||
+                                                optionUser.name ||
+                                                optionUser.username ||
+                                                ''
+                                            );
+                                        } else {
+                                            return option.entity.name;
+                                        }
+                                    }}
+                                    filterOptions={(options, { inputValue }) =>
+                                        options.filter(
+                                            (option: IAccessOption) => {
+                                                if (
+                                                    option.type ===
+                                                        ENTITY_TYPE.USER ||
+                                                    option.type ===
+                                                        ENTITY_TYPE.SERVICE_ACCOUNT
+                                                ) {
+                                                    const optionUser =
+                                                        option.entity as IUser;
+                                                    return (
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.email,
+                                                        ) ||
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.name,
+                                                        ) ||
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.username,
+                                                        )
+                                                    );
+                                                }
+                                                return caseInsensitiveSearch(
+                                                    inputValue,
+                                                    option.entity.name,
+                                                );
+                                            },
+                                        )
+                                    }
+                                    isOptionEqualToValue={(option, value) =>
+                                        option.type === value.type &&
+                                        option.entity.id === value.entity.id
+                                    }
+                                    renderInput={(params) => (
+                                        <TextField
+                                            {...params}
+                                            label={capitalize(entityType)}
+                                        />
+                                    )}
+                                />
+                            ) : (
+                                <Autocomplete
+                                    data-testid={PA_USERS_GROUPS_ID}
+                                    size='small'
+                                    multiple
+                                    openOnFocus
+                                    limitTags={10}
+                                    disableCloseOnSelect
+                                    disabled={edit}
+                                    value={selectedOptions}
+                                    getOptionDisabled={(option) => {
+                                        if (option.type === ENTITY_TYPE.GROUP) {
+                                            const optionGroup =
+                                                option.entity as IGroup;
+                                            return Boolean(
+                                                optionGroup.rootRole,
+                                            );
+                                        }
+                                        return false;
+                                    }}
+                                    onChange={(event, newValue, reason) => {
+                                        if (
+                                            event.type === 'keydown' &&
+                                            (event as React.KeyboardEvent)
+                                                .key === 'Backspace' &&
+                                            reason === 'removeOption'
+                                        ) {
+                                            return;
+                                        }
+                                        setSelectedOptions(newValue);
+                                    }}
+                                    options={options}
+                                    groupBy={(option) => option.type}
+                                    renderOption={(
+                                        props,
+                                        option,
+                                        { selected },
+                                    ) => renderOption(props, option, selected)}
+                                    getOptionLabel={(option: IAccessOption) => {
+                                        if (
+                                            option.type === ENTITY_TYPE.USER ||
+                                            option.type ===
+                                                ENTITY_TYPE.SERVICE_ACCOUNT
+                                        ) {
+                                            const optionUser =
+                                                option.entity as IUser;
+                                            return (
+                                                optionUser.email ||
+                                                optionUser.name ||
+                                                optionUser.username ||
+                                                ''
+                                            );
+                                        } else {
+                                            return option.entity.name;
+                                        }
+                                    }}
+                                    filterOptions={(options, { inputValue }) =>
+                                        options.filter(
+                                            (option: IAccessOption) => {
+                                                if (
+                                                    option.type ===
+                                                        ENTITY_TYPE.USER ||
+                                                    option.type ===
+                                                        ENTITY_TYPE.SERVICE_ACCOUNT
+                                                ) {
+                                                    const optionUser =
+                                                        option.entity as IUser;
+                                                    return (
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.email,
+                                                        ) ||
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.name,
+                                                        ) ||
+                                                        caseInsensitiveSearch(
+                                                            inputValue,
+                                                            optionUser.username,
+                                                        )
+                                                    );
+                                                }
+                                                return caseInsensitiveSearch(
+                                                    inputValue,
+                                                    option.entity.name,
+                                                );
+                                            },
+                                        )
+                                    }
+                                    isOptionEqualToValue={(option, value) =>
+                                        option.type === value.type &&
+                                        option.entity.id === value.entity.id
+                                    }
+                                    renderInput={(params) => (
+                                        <TextField
+                                            {...params}
+                                            label={capitalize(entityType)}
+                                        />
+                                    )}
+                                />
+                            )}
                         </StyledAutocompleteWrapper>
                         <StyledInputDescription>
                             Select the role to assign for this project
